+++ packages/perseus/src/widgets/interactive-graphs/graphs/components/angle-indicators.test.ts
@@ -1,8 +1,9 @@
-import {findAngle} from "../../math";
+import {getClockwiseAngle} from "../../math";
 
 import {shouldDrawArcOutside} from "./angle-indicators";
 
 import type {CollinearTuple} from "../../../../perseus-types";
+import type {Coord} from "@khanacademy/perseus";
 import type {vec, Interval} from "mafs";
 
 describe("shouldDrawArcOutside", () => {
@@ -94,10 +95,12 @@ describe("shouldDrawArcOutside", () => {
         ).toBe(true);
     });
 
+    // TODO: (third) Move this test to the math package
     it("should correctly calculate the angle for the given coordinates", () => {
         const point1 = [2, 2] as vec.Vector2;
         const point2 = [2, 0] as vec.Vector2;
         const vertex = [0, 0] as vec.Vector2;
-        expect(findAngle(point1, point2, vertex)).toBe(45);
+        const coords: [Coord, Coord, Coord] = [point1, vertex, point2];
+        expect(getClockwiseAngle(coords)).toBe(45);
     });
 });