+++ packages/pure-markdown/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/pure-markdown
 
+## 0.3.10
+
+### Patch Changes
+
+-   Updated dependencies [[`d4f4e2be1`](https://github.com/Khan/perseus/commit/d4f4e2be1408c4531a146bcd496344a629d90bd1)]:
+    -   @khanacademy/perseus-core@1.5.2
+    -   @khanacademy/simple-markdown@0.13.3
+
 ## 0.3.9
 
 ### Patch Changes